+++ eras/dijkstra/impl/testlib/Test/Cardano/Ledger/Dijkstra/Imp/EntitiesSpec.hs
@@ -13,27 +13,92 @@ import Cardano.Ledger.Address
 import Cardano.Ledger.BaseTypes
 import Cardano.Ledger.Coin (Coin (..))
 import Cardano.Ledger.Credential (Credential (..))
-import Cardano.Ledger.DRep (DRep (..))
 import Cardano.Ledger.Dijkstra.Core
 import Cardano.Ledger.Dijkstra.Rules (
-  DijkstraUtxoPredFailure (..),
   EntitiesPredFailure (..),
   SubEntitiesPredFailure (..),
  )
 import Cardano.Ledger.Dijkstra.Scripts (AccountBalanceInterval (..), AccountBalanceIntervals (..))
-import Cardano.Ledger.Plutus
 import Cardano.Ledger.Val (Val (..))
-import qualified Data.Map.NonEmpty as NE
-import qualified Data.Map.Strict as Map
+import qualified Data.Map.NonEmpty as NEM
 import Data.Maybe (fromJust)
+import qualified Data.OMap.Strict as OMap
 import qualified Data.Set.NonEmpty as NES
-import Lens.Micro ((&), (.~))
+import Lens.Micro
 import Test.Cardano.Ledger.Dijkstra.ImpTest
 import Test.Cardano.Ledger.Imp.Common
-import Test.Cardano.Ledger.Plutus.Examples (alwaysSucceedsWithDatum)
 
 spec :: forall era. DijkstraEraImp era => SpecWith (ImpInit (LedgerSpec era))
 spec = describe "ENTITIES" $ do
+  it "Batch with successful withdrawals and direct deposits" $ do
+    modifyPParams $ ppGovActionLifetimeL .~ EpochInterval 2
+    (acc1, balance1, kh1) <- setupAccountAddress
+    (acc2, balance2, kh2) <- setupAccountAddress
+    (acc3, balance3, kh3) <- setupAccountAddress
+
+    let depositAmount = Coin 50
+        partialWithdrawal = balance3 <-> Coin 10
+        subDeposit =
+          mkBasicTx $
+            mkBasicTxBody
+              & directDepositsTxBodyL .~ DirectDeposits [(acc1, depositAmount)]
+        subWithdraw =
+          mkBasicTx $
+            mkBasicTxBody
+              & withdrawalsTxBodyL .~ Withdrawals [(acc2, balance2)]
+        topTx =
+          mkBasicTx $
+            mkBasicTxBody
+              & withdrawalsTxBodyL .~ Withdrawals [(acc3, partialWithdrawal)]
+              & subTransactionsTxBodyL .~ [subDeposit, subWithdraw]
+    submitTx_ topTx
+
+    finalBalance1 <- getBalance (KeyHashObj kh1)
+    finalBalance2 <- getBalance (KeyHashObj kh2)
+    finalBalanceD <- getBalance (KeyHashObj kh3)
+
+    finalBalance1 `shouldBe` balance1 <+> depositAmount
+    finalBalance2 `shouldBe` mempty
+    finalBalanceD `shouldBe` (balance3 <-> partialWithdrawal)
+
+  it "Partial withdrawals" $ do
+    modifyPParams $ ppGovActionLifetimeL .~ EpochInterval 2
+
+    (account1, balance1, kh1) <- setupAccountAddress
+    (account2, balance2, kh2) <- setupAccountAddress
+    lessThanBalance1 <- Coin <$> choose (1, unCoin balance1 - 1)
+    atMostBalance2 <- Coin <$> choose (1, unCoin balance2)
+    let tx =
+          mkTxWithBatchWithdrawals
+            (Withdrawals [(account1, lessThanBalance1)])
+            [Withdrawals [(account2, atMostBalance2)]]
+    submitTx_ tx
+    getBalance (KeyHashObj kh1) `shouldReturn` (balance1 <-> lessThanBalance1)
+    getBalance (KeyHashObj kh2) `shouldReturn` (balance2 <-> atMostBalance2)
+
+    -- restore balances, to test legacy mode
+    submitTx_ $
+      mkBasicTx $
+        mkBasicTxBody
+          & directDepositsTxBodyL .~ DirectDeposits [(account1, lessThanBalance1), (account2, atMostBalance2)]
+    legacyTx <- switchTxToLegacyMode tx
+    submitFailingTx
+      legacyTx
+      [ injectFailure . WithdrawalAmountsInexactInLegacyMode @era $
+          NEM.singleton account1 $
+            Mismatch lessThanBalance1 balance1
+      ]
+
+    -- drain top withdrawal
+    submitTx_
+      =<< switchTxToLegacyMode
+        ( mkTxWithBatchWithdrawals
+            (Withdrawals [(account1, balance1)])
+            [Withdrawals [(account2, atMostBalance2)]]
+        )
+    getBalance (KeyHashObj kh1) `shouldReturn` zero
+    getBalance (KeyHashObj kh2) `shouldReturn` (balance2 <-> atMostBalance2)
+
   it "Withdrawals from an unregistered staking address" $ do
     modifyPParams $ ppGovActionLifetimeL .~ EpochInterval 2
 
@@ -47,11 +112,7 @@ spec = describe "ENTITIES" $ do
           & withdrawalsTxBodyL .~ Withdrawals [(account1, amountX), (account2, zero)]
     submitFailingTx
       (mkBasicTx txBody)
-      [ injectFailure $
-          WithdrawalsExceedAccountBalance @era $
-            NE.singleton account1 $
-              Mismatch amountX mempty
-      , injectFailure . MissingAccountsInWithdrawals @era $
+      [ injectFailure . WithdrawalAccountsMissingFromOriginal @era $
           Withdrawals [(account1, amountX), (account2, zero)]
       ]
 
@@ -63,23 +124,15 @@ spec = describe "ENTITIES" $ do
               & withdrawalsTxBodyL .~ Withdrawals [(account3, amountY)]
     submitFailingTx
       (mkBasicTx $ txBody & subTransactionsTxBodyL .~ [mkBasicTx txBody, subTxOnlyWithdrawal])
-      [ injectFailure $
-          WithdrawalsExceedAccountBalance @era $
-            fromJust $
-              NE.fromMap $
-                Map.fromList
-                  [ (account1, Mismatch (amountX <> amountX) mempty)
-                  , (account3, Mismatch amountY mempty)
-                  ]
-      , injectFailure . MissingAccountsInWithdrawals @era $
+      [ injectFailure . WithdrawalAccountsMissingFromOriginal @era $
           Withdrawals [(account1, amountX), (account2, zero)]
-      , injectFailure . SubMissingOriginalAccountsInWithdrawals @era $
+      , injectFailure . SubWithdrawalAccountsMissingFromOriginal @era $
           Withdrawals [(account1, amountX), (account2, zero)]
-      , injectFailure . SubMissingAccountsInWithdrawals @era $
+      , injectFailure . SubWithdrawalAccountsMissing @era $
           Withdrawals [(account1, amountX), (account2, zero)]
-      , injectFailure . SubMissingOriginalAccountsInWithdrawals @era $
+      , injectFailure . SubWithdrawalAccountsMissingFromOriginal @era $
           Withdrawals [(account3, amountY)]
-      , injectFailure . SubMissingAccountsInWithdrawals @era $
+      , injectFailure . SubWithdrawalAccountsMissing @era $
           Withdrawals [(account3, amountY)]
       ]
 
@@ -87,107 +140,231 @@ spec = describe "ENTITIES" $ do
     account <- freshKeyHash >>= getAccountAddressFor . KeyHashObj
     amountX <- Coin . getPositive <$> arbitrary
     let
-    let
       txBody :: forall l. Typeable l => TxBody l era
       txBody = mkBasicTxBody & directDepositsTxBodyL .~ DirectDeposits [(account, amountX)]
     submitFailingTx
       (mkBasicTx txBody)
-      [ injectFailure . MissingAccountsInDirectDeposits @era $
+      [ injectFailure . DirectDepositAccountsMissing @era $
           DirectDeposits [(account, amountX)]
       ]
 
     account2 <- freshKeyHash >>= getAccountAddressFor . KeyHashObj
     amountY <- Coin . getPositive <$> arbitrary
     amountZ <- Coin . getPositive <$> arbitrary
-    let
     let subTxOnlyDirectDeposit =
           mkBasicTx $
             mkBasicTxBody & directDepositsTxBodyL .~ DirectDeposits [(account, amountY), (account2, amountZ)]
     submitFailingTx
       (mkBasicTx $ txBody & subTransactionsTxBodyL .~ [mkBasicTx txBody, subTxOnlyDirectDeposit])
-      [ injectFailure . MissingAccountsInDirectDeposits @era $
+      [ injectFailure . DirectDepositAccountsMissing @era $
           DirectDeposits [(account, amountX)]
-      , injectFailure . SubMissingAccountsInDirectDeposits @era $
+      , injectFailure . SubDirectDepositAccountsMissing @era $
           DirectDeposits [(account, amountX)]
-      , injectFailure . SubMissingAccountsInDirectDeposits @era $
+      , injectFailure . SubDirectDepositAccountsMissing @era $
           DirectDeposits [(account, amountY), (account2, amountZ)]
       ]
-
-  it "Withdrawals of the wrong amount" $ do
-    modifyPParams $ ppGovActionLifetimeL .~ EpochInterval 2
-
-    (accountAddress1, reward1, stakeKey1) <- setupAccountAddress
-    (accountAddress2, reward2, stakeKey2) <- setupAccountAddress
-    void $ delegateToDRep (KeyHashObj stakeKey1) (Coin 1_000_000) DRepAlwaysAbstain
-    void $ delegateToDRep (KeyHashObj stakeKey2) (Coin 1_000_000) DRepAlwaysAbstain
-    submitFailingTx
-      ( mkBasicTx $
-          mkBasicTxBody
-            & withdrawalsTxBodyL
-              .~ Withdrawals
-                [ (accountAddress1, reward1 <+> Coin 1)
-                , (accountAddress2, reward2)
-                ]
-      )
-      [ injectFailure $
-          WithdrawalsExceedAccountBalance @era $
-            NE.singleton accountAddress1 $
-              Mismatch (reward1 <+> Coin 1) reward1
-      , injectFailure $
-          ExceededBalancesInWithdrawals @era $
-            NE.singleton accountAddress1 $
-              Mismatch (reward1 <+> Coin 1) reward1
-      ]
-
-    -- in legacy mode, we produce `IncompleteWithdrawals` failure
-    txIn <- produceScript . hashPlutusScript $ alwaysSucceedsWithDatum SPlutusV2
-    submitFailingTx
-      ( mkBasicTx $
-          mkBasicTxBody
-            & withdrawalsTxBodyL
-              .~ Withdrawals
-                [(accountAddress1, zero)]
-            & inputsTxBodyL .~ [txIn]
-      )
-      [ injectFailure . IncompleteWithdrawals @era $
-          NE.singleton accountAddress1 $
-            Mismatch zero reward1
-      ]
-
-    submitTx_ $
-      mkBasicTx $
-        mkBasicTxBody
-          & withdrawalsTxBodyL
-            .~ Withdrawals
-              [(accountAddress1, zero)]
 
   it "Withdrawals and direct deposits with wrong network id" $ do
     stakeKey <- freshKeyHash
     accountAddress <- registerStakeCredential (KeyHashObj stakeKey)
     let wrongNetworkAccount = accountAddress & accountAddressNetworkIdL .~ Mainnet
+    let dd = DirectDeposits [(wrongNetworkAccount, Coin 50)]
     let txBody :: forall l. Typeable l => TxBody l era
         txBody =
           mkBasicTxBody
             & withdrawalsTxBodyL
               .~ Withdrawals [(wrongNetworkAccount, mempty)]
-            & directDepositsTxBodyL
-              .~ DirectDeposits [(wrongNetworkAccount, Coin 50)]
+            & directDepositsTxBodyL .~ dd
 
     submitFailingTx
       (mkBasicTx txBody)
-      [ injectFailure . WrongNetworkInWithdrawals @era Testnet $ NES.singleton wrongNetworkAccount
-      , injectFailure . WrongNetworkInDirectDeposits @era Testnet $ NES.singleton wrongNetworkAccount
-      , injectFailure . MissingAccountsInWithdrawals @era $ Withdrawals [(wrongNetworkAccount, mempty)]
+      [ injectFailure . WithdrawalAddressesWithWrongNetwork @era Testnet $ NES.singleton wrongNetworkAccount
+      , injectFailure . DirectDepositAddressesWithWrongNetwork @era Testnet $
+          NES.singleton wrongNetworkAccount
+      , injectFailure . WithdrawalAccountsMissingFromOriginal @era $
+          Withdrawals [(wrongNetworkAccount, mempty)]
+      , injectFailure . DirectDepositAccountsMissing @era $ dd
       ]
 
     submitFailingTx
       (mkBasicTx $ txBody & subTransactionsTxBodyL .~ [mkBasicTx txBody])
-      [ injectFailure . WrongNetworkInWithdrawals @era Testnet $ NES.singleton wrongNetworkAccount
-      , injectFailure . WrongNetworkInDirectDeposits @era Testnet $ NES.singleton wrongNetworkAccount
-      , injectFailure . MissingAccountsInWithdrawals @era $ Withdrawals [(wrongNetworkAccount, mempty)]
-      , injectFailure . SubWrongNetworkInWithdrawals @era Testnet $ NES.singleton wrongNetworkAccount
-      , injectFailure . SubWrongNetworkInDirectDeposits @era Testnet $ NES.singleton wrongNetworkAccount
+      [ injectFailure . WithdrawalAddressesWithWrongNetwork @era Testnet $
+          NES.singleton wrongNetworkAccount
+      , injectFailure . DirectDepositAddressesWithWrongNetwork @era Testnet $
+          NES.singleton wrongNetworkAccount
+      , injectFailure . WithdrawalAccountsMissingFromOriginal @era $
+          Withdrawals [(wrongNetworkAccount, mempty)]
+      , injectFailure . DirectDepositAccountsMissing @era $ dd
+      , injectFailure . SubWithdrawalAddressesWithWrongNetwork @era Testnet $
+          NES.singleton wrongNetworkAccount
+      , injectFailure . SubDirectDepositAddressesWithWrongNetwork @era Testnet $
+          NES.singleton wrongNetworkAccount
+      , injectFailure . SubWithdrawalAccountsMissingFromOriginal @era $
+          Withdrawals [(wrongNetworkAccount, mempty)]
+      , injectFailure . SubWithdrawalAccountsMissing @era $
+          Withdrawals [(wrongNetworkAccount, mempty)]
+      , injectFailure . SubDirectDepositAccountsMissing @era $ dd
       ]
+
+  it "Aggregate of top and sub withdrawals exceeds account balance" $ do
+    modifyPParams $ ppGovActionLifetimeL .~ EpochInterval 2
+    (account, balance, _) <- setupAccountAddress
+    (topAmount, subAmount) <- genCoinPairExceeding balance
+    let tx =
+          mkTxWithBatchWithdrawals
+            (Withdrawals [(account, topAmount)])
+            [Withdrawals [(account, subAmount)]]
+    submitFailingTx
+      tx
+      [ injectFailure $
+          WithdrawalAmountsExceedingOriginalBalance @era $
+            fromJust $
+              NEM.fromMap [(account, Mismatch (topAmount <+> subAmount) balance)]
+      ]
+    -- legacy mode
+    legacyTx <- switchTxToLegacyMode tx
+    submitFailingTx
+      legacyTx
+      [ injectFailure . WithdrawalAmountsInexactInLegacyMode @era $
+          NEM.singleton account $
+            Mismatch topAmount (balance <-> subAmount)
+      ]
+
+  it "Aggregate of sub withdrawals exceeds account balance" $ do
+    modifyPParams $ ppGovActionLifetimeL .~ EpochInterval 2
+    (account, balance, _) <- setupAccountAddress
+    (subAmount1, subAmount2) <- genCoinPairExceeding balance
+    (subAmount1 <+> subAmount2) `shouldSatisfy` (> balance)
+
+    let tx =
+          mkTxWithBatchWithdrawals
+            (Withdrawals [(account, zero)])
+            [Withdrawals [(account, subAmount1)], Withdrawals [(account, subAmount2)]]
+    submitFailingTx
+      tx
+      [ injectFailure $
+          WithdrawalAmountsExceedingOriginalBalance @era $
+            fromJust $
+              NEM.fromMap [(account, Mismatch (subAmount1 <+> subAmount2) balance)]
+      ]
+    legacyTx <- switchTxToLegacyMode tx
+    submitFailingTx
+      legacyTx
+      [ injectFailure $
+          WithdrawalAmountsExceedingOriginalBalance @era $
+            fromJust $
+              NEM.fromMap [(account, Mismatch (subAmount1 <+> subAmount2) balance)]
+      ]
+
+  it "Individual withdrawal exceeds account balance" $ do
+    modifyPParams $ ppGovActionLifetimeL .~ EpochInterval 2
+    (account, balance, _) <- setupAccountAddress
+    atMostBalance <- Coin <$> choose (1, unCoin balance)
+    moreThanBalance <- (balance <+>) . Coin . getPositive <$> arbitrary
+
+    -- A sub-transaction overdraws
+    let subTxOverdraws =
+          mkTxWithBatchWithdrawals
+            (Withdrawals [(account, atMostBalance)])
+            [Withdrawals [(account, moreThanBalance)]]
+    submitFailingTx
+      subTxOverdraws
+      [ injectFailure $
+          WithdrawalAmountsExceedingOriginalBalance @era $
+            fromJust $
+              NEM.fromMap [(account, Mismatch (atMostBalance <+> moreThanBalance) balance)]
+      ]
+
+    legacySubTxOverdraws <- switchTxToLegacyMode subTxOverdraws
+
+    submitFailingTx
+      legacySubTxOverdraws
+      [ injectFailure $
+          WithdrawalAmountsExceedingOriginalBalance @era $
+            fromJust $
+              NEM.fromMap [(account, Mismatch moreThanBalance balance)]
+      ]
+
+    -- The top transaction overdraws
+    let topTxOverdraws =
+          mkTxWithBatchWithdrawals
+            (Withdrawals [(account, moreThanBalance)])
+            [Withdrawals [(account, atMostBalance)]]
+    submitFailingTx
+      topTxOverdraws
+      [ injectFailure $
+          WithdrawalAmountsExceedingOriginalBalance @era $
+            fromJust $
+              NEM.fromMap
+                [(account, Mismatch (atMostBalance <+> moreThanBalance) balance)]
+      ]
+    legacyTopTxOverdraws <- switchTxToLegacyMode topTxOverdraws
+    submitFailingTx
+      legacyTopTxOverdraws
+      [ injectFailure . WithdrawalAmountsInexactInLegacyMode @era $
+          NEM.singleton account $
+            Mismatch moreThanBalance (balance <-> atMostBalance)
+      ]
+
+  it "Direct deposits cannot fund withdrawals in subsequent sub-transactions" $ do
+    account <- registerStakeCredential . KeyHashObj =<< freshKeyHash
+    depositAmount <- Coin . getPositive <$> arbitrary
+    let subDeposit =
+          mkBasicTx $
+            mkBasicTxBody
+              & directDepositsTxBodyL
+                .~ DirectDeposits [(account, depositAmount)]
+        subWithdraw =
+          mkBasicTx $
+            mkBasicTxBody
+              & withdrawalsTxBodyL
+                .~ Withdrawals [(account, depositAmount)]
+        tx =
+          mkBasicTx $
+            mkBasicTxBody
+              & subTransactionsTxBodyL .~ [subDeposit, subWithdraw]
+
+    submitFailingTx
+      tx
+      [ injectFailure $
+          WithdrawalAmountsExceedingOriginalBalance @era $
+            fromJust $
+              NEM.fromMap [(account, Mismatch depositAmount zero)]
+      ]
+
+    legacyTx <- switchTxToLegacyMode tx
+    submitFailingTx
+      legacyTx
+      [ injectFailure $
+          WithdrawalAmountsExceedingOriginalBalance @era $
+            fromJust $
+              NEM.fromMap [(account, Mismatch depositAmount zero)]
+      ]
+
+  it "Top transaction can drain an account funded by a sub-transaction direct deposit, in legacy mode" $ do
+    account <- registerStakeCredential . KeyHashObj =<< freshKeyHash
+    depositAmount <- Coin . getPositive <$> arbitrary
+    let subDeposit =
+          mkBasicTx $
+            mkBasicTxBody
+              & directDepositsTxBodyL .~ DirectDeposits [(account, depositAmount)]
+        tx =
+          mkBasicTx $
+            mkBasicTxBody
+              & withdrawalsTxBodyL .~ Withdrawals [(account, depositAmount)]
+              & subTransactionsTxBodyL .~ [subDeposit]
+
+    submitFailingTx
+      tx
+      [ injectFailure $
+          WithdrawalAmountsExceedingOriginalBalance @era $
+            fromJust $
+              NEM.fromMap [(account, Mismatch depositAmount zero)]
+      ]
+
+    legacyTx <- switchTxToLegacyMode tx
+    submitTx_ legacyTx
+    getBalance (account ^. accountAddressCredentialL) `shouldReturn` zero
 
   describe "Account balance intervals" $ do
     it "Account balance intervals for the top-level transaction" $
@@ -240,10 +417,10 @@ spec = describe "ENTITIES" $ do
         [ injectFailure $
             WrongNetworkInAccountBalanceIntervals @era Testnet (NES.singleton onWrongNetwork)
         , injectFailure $
-            MissingAccountsInAccountBalanceIntervals @era (NE.singleton unregistered violated)
+            MissingAccountsInAccountBalanceIntervals @era (NEM.singleton unregistered violated)
         , injectFailure $
             BalancesOutsideAccountBalanceIntervals @era
-              (NE.singleton accountAddr (balance, violated))
+              (NEM.singleton accountAddr (balance, violated))
         ]
 
     it
@@ -263,7 +440,7 @@ spec = describe "ENTITIES" $ do
                 (txWith modifyBody interval)
                 [ injectFailure $
                     BalancesOutsideAccountBalanceIntervals @era
-                      (NE.singleton accountAddr (balance, interval))
+                      (NEM.singleton accountAddr (balance, interval))
                 ]
         expectOutside drains (AccountBalanceExact zero)
         expectOutside deposits (AccountBalanceExact (balance <+> deposit))
@@ -280,7 +457,7 @@ spec = describe "ENTITIES" $ do
           )
           [ injectFailure $
               SubBalancesOutsideAccountBalanceIntervals @era
-                (NE.singleton accountAddr (balance, AccountBalanceExact zero))
+                (NEM.singleton accountAddr (balance, AccountBalanceExact zero))
           ]
         submitTx_ $ txWith drains (AccountBalanceExact balance)
 
@@ -296,7 +473,7 @@ spec = describe "ENTITIES" $ do
               (withInterval interval)
               [ injectFailure $
                   BalancesOutsideAccountBalanceIntervals @era
-                    (NE.singleton accountAddr (balance, interval))
+                    (NEM.singleton accountAddr (balance, interval))
               ]
       intervalHolds $ AccountBalanceExact balance
       intervalViolated $ AccountBalanceExact (balance <+> Coin 1)
@@ -328,10 +505,10 @@ spec = describe "ENTITIES" $ do
         submitFailingTx
           (txWithIntervals drained original)
           [ injectFailure $
-              BalancesOutsideAccountBalanceIntervals @era (NE.singleton accountAddr (zero, original))
+              BalancesOutsideAccountBalanceIntervals @era (NEM.singleton accountAddr (zero, original))
           , injectFailure $
               BalancesOutsideStartingAccountBalanceIntervals @era
-                (NE.singleton accountAddr (balance, drained))
+                (NEM.singleton accountAddr (balance, drained))
           ]
         submitTx_ $ txWithIntervals original drained
   where
@@ -343,6 +520,21 @@ spec = describe "ENTITIES" $ do
       submitAndExpireProposalToMakeReward cred
       b <- getBalance cred
       pure (ra, b, kh)
+
+    mkTxWithBatchWithdrawals :: Withdrawals -> [Withdrawals] -> Tx TopTx era
+    mkTxWithBatchWithdrawals topWdrls subs =
+      mkBasicTx $
+        mkBasicTxBody
+          & withdrawalsTxBodyL .~ topWdrls
+          & subTransactionsTxBodyL .~ OMap.fromFoldable (fmap mkSubTx subs)
+      where
+        mkSubTx :: Withdrawals -> Tx SubTx era
+        mkSubTx w = mkBasicTx (mkBasicTxBody & withdrawalsTxBodyL .~ w)
+
+    genCoinPairExceeding (Coin maxSum) = do
+      a <- choose (1, maxSum)
+      b <- choose (maxSum - a + 1, maxSum)
+      pure (Coin a, Coin b)
 
     unregisteredAccount :: ImpTestM era AccountAddress
     unregisteredAccount = freshKeyHash >>= getAccountAddressFor . KeyHashObj
@@ -358,8 +550,8 @@ spec = describe "ENTITIES" $ do
     accountBalanceIntervalCases ::
       (AccountBalanceIntervals era -> t era -> ImpTestM era ()) ->
       (Network -> NES.NonEmptySet AccountAddress -> t era) ->
-      (NE.NonEmptyMap AccountAddress (AccountBalanceInterval era) -> t era) ->
-      (NE.NonEmptyMap AccountAddress (Coin, AccountBalanceInterval era) -> t era) ->
+      (NEM.NonEmptyMap AccountAddress (AccountBalanceInterval era) -> t era) ->
+      (NEM.NonEmptyMap AccountAddress (Coin, AccountBalanceInterval era) -> t era) ->
       ImpTestM era ()
     accountBalanceIntervalCases submitFailing mkWrongNetwork mkMissingAccounts mkBalancesOutside = do
       (accountAddr, balance, _) <- setupAccountAddress
@@ -371,7 +563,7 @@ spec = describe "ENTITIES" $ do
       unregistered <- unregisteredAccount
       submitFailing
         (AccountBalanceIntervals [(unregistered, violated)])
-        (mkMissingAccounts (NE.singleton unregistered violated))
+        (mkMissingAccounts (NEM.singleton unregistered violated))
       submitFailing
         (AccountBalanceIntervals [(accountAddr, violated)])
-        (mkBalancesOutside (NE.singleton accountAddr (balance, violated)))
+        (mkBalancesOutside (NEM.singleton accountAddr (balance, violated)))